+++ common/src/main/java/dev/aika/smsn/SMSN.java
@@ -1,6 +1,7 @@
 package dev.aika.smsn;
 
 import dev.aika.smsn.config.SMSNConfig;
+import dev.aika.smsn.mixin.ModMixinManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -10,6 +11,7 @@ public final class SMSN {
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_NAME);
 
     public static SMSNConfig CONFIG = new SMSNConfig();
+    public static final ModMixinManager MixinManager = new ModMixinManager();
 
     public static void init() {
         CONFIG = SMSNConfig.load();