+++ common/src/main/java/dev/aika/smsn/client/gui/components/SwitchBuilder.java
@@ -17,22 +17,18 @@ public final class SwitchBuilder extends AbstractComponentBuilder<Boolean> {
     public static final String DefaultCheckedText = "true";
     public static final String DefaultUncheckedText = "false";
 
-    SwitchBuilder(ConfigEntryBuilder entryBuilder, Object object, Field field) {
-        super(entryBuilder, object);
-
-        this.field = field;
+    public SwitchBuilder(ConfigEntryBuilder entryBuilder, Object object, Field field) {
+        super(entryBuilder, object, field);
         this.annotation = field.getAnnotation(Components.Switch.class);
     }
 
     @Override
     public AbstractConfigListEntry<Boolean> build() {
         BooleanToggleBuilder builder = entryBuilder
-                .startBooleanToggle(
-                        Component.translatable(String.format("%s.%s", translatableKeyPrefix(), field.getName())),
-                        getValue())
+                .startBooleanToggle(fieldNameKey(), getValue())
                 .setSaveConsumer(this::setValue)
                 .setYesNoTextSupplier(this::yesNoTextSupplier);
-        if (defaultObject != null) builder = builder.setDefaultValue(getDefaultValue());
+        fieldBuilderInit(builder);
         return builder.build();
     }