+++ fabric/src/main/java/dev/aika/smsn/fabric/compat/ClothConfigCompat.java
@@ -3,8 +3,9 @@ package dev.aika.smsn.fabric.compat;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import dev.aika.smsn.SMSN;
+import dev.aika.smsn.api.ModPlatform;
 import dev.aika.smsn.client.gui.MissingClothConfigScreen;
-import dev.aika.smsn.fabric.SMSNPlatformImpl;
+import dev.aika.smsn.compat.ClothConfigScreen;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 
@@ -12,10 +13,12 @@ import net.fabricmc.api.Environment;
 public class ClothConfigCompat implements ModMenuApi {
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return parent -> {
-            if (SMSNPlatformImpl.isModLoaded("cloth-config2"))
-                return dev.aika.smsn.compat.ClothConfigCompat.ConfigScreen(SMSN.CONFIG, parent);
-            else return new MissingClothConfigScreen(parent);
-        };
+        return parent ->
+                ModPlatform.isModLoaded("cloth-config2") ?
+                        ClothConfigScreen.builder()
+                                .setParent(parent).setModId(SMSN.MOD_ID)
+                                .setConfig(SMSN.CONFIG).setMixinManager(SMSN.MixinManager)
+                                .build()
+                        : new MissingClothConfigScreen(parent);
     }
 }